+++ src/GraphicsEngine/Vulkan/VulkanRenderEngineRasterizer.hpp
@@ -62,7 +62,7 @@ public:
                 vkCmdBindIndexBuffer(commandBufferManager.commandBuffers[imageIndex], asset->indexBuffer.buffer, 0, VK_INDEX_TYPE_UINT32);
                 vkCmdBindDescriptorSets(commandBufferManager.commandBuffers[imageIndex], VK_PIPELINE_BIND_POINT_GRAPHICS, asset->pipelineManagers[0].pipelineLayout, 0, 1, &asset->pipelineManagers[0].descriptorSet, 0, nullptr);
                 vkCmdBindPipeline(commandBufferManager.commandBuffers[imageIndex], VK_PIPELINE_BIND_POINT_GRAPHICS, asset->pipelineManagers[0].pipeline);
-                vkCmdDrawIndexed(commandBufferManager.commandBuffers[imageIndex], asset->indices.size(), 1, 0, 0, 0);
+                vkCmdDrawIndexed(commandBufferManager.commandBuffers[imageIndex], static_cast<uint32_t>(asset->indices.size()), 1, 0, 0, 0);
             }
         }
         vkCmdEndRenderPass(commandBufferManager.commandBuffers[imageIndex]);
